Now the defination of the sixth column is exactly same as Plink. Now the style of the commands is GNU, and supports one character long, such as "--snp-list".

+++ UnifiedFramework/src/family/pedigree/design/hierarchy/ChenInterface.java
@@ -15,7 +15,7 @@ public interface ChenInterface {
 
 	public void setSeed(long s);
 	public void recoverSeed();
-	public byte[] getStatus();
+	public double[] getStatus();
 	public void getPermutedScore(boolean nested);
 	public void RecoverScore();
 	public MapFile getMapFile();
